Update v1_TeslaMateAPICarsCharges.go
It's possible to have incomplete charges with missing EndDate. These aren't always valid objects either, but they shouldn't prevent the Charges endpoint from providing data.
I believe this can happen when TeslaMate is offline when the charging cycle ends. I am not sure if it's also possible if the vehicle is actively charging.

Feel free to edit this PR, not sure that an empty string is the best option.