Update Charges to show how this could work
Only added surface level implementation to one place as a proof of concept. Theoretically, we would sprinkle `TeslaMateAPIHandleErrorResponse` across the endpoint files instead of waiting until the end. Then we can forget `ValidResponse` entirely, and know that it is successful if we reach the bottom.

The final goal here it to remove instances of log.Fatal(err) where the app can no longer handle an error response. When it crashes, the connection is dropped and the user isn't provided an error. This lays the foundation to fix it.