Provide Access to Incomplete Charges
While these are usually charges that failed, they also include charges that are in progress. The client should be able to see these Charges and decide how to proceed. These charges were, at one point, real charges with relevant data.

Reasons for failure:
- Charging started, TeslaMate server or Tesla API went offline
- Charging in progress (not actually failure, just incomplete charge data)

This will also allow the user to know about the charge's existence and query (by ID) the Charge Details. Since the database stored the many charge detail objects, the user may be able to process some additional information about the charge. Especially if the charge is in progress.

Example:
1. Charge is in progress.
2. Receive basic charge data, like start date and charge ID. 
3. Query Charge Details to info like: start date, power over time, etc.

However, even with access to this additional data that does exist, the user still will not be able to see data from when the charge went offline. Or in the case the charge is still occurring, the user can't see charge details that haven't happened yet.

Therefore,
Charge Details require an additional parameter `isCompleted:BOOL` so the user knows if an object is finalized.

**The user will have to determine whether the charge is still in progress or corrupted, by seeing if there are any recent charges. If it's in progress, the user should consider refreshing that charge object in the future.**

I chose to add this as a separate object on `Data` so that it will not affect existing API users. This will ensure incomplete charges are not treated as completed charges.